Makes the license valid
Just letting you know, the license you currently have for this project isn't "valid". That being; right now, even with your current license file, your work is not under an actual license.

The minimum you have to do to make it legally valid, would be to change your license file to have the `.md` extension, and from there convert the links and image to markdown format. I have made a pull request, doing this for you.